+++ tests/utils/test_check_column.py
@@ -1,5 +1,4 @@
 import pytest
-from hypothesis import given
 
 from janitor.utils import check_column
 
@@ -7,7 +6,7 @@ from janitor.utils import check_column
 @pytest.mark.utils
 def test_check_column(dataframe):
     """
-    rename_column should return true if column exist
+    check_column should return if column exist
     """
     assert check_column(dataframe, ["a"]) is None
 
@@ -15,7 +14,25 @@ def test_check_column(dataframe):
 @pytest.mark.utils
 def test_check_column_absent_column(dataframe):
     """
-    rename_column should raise an error if the column is absent.
+    check_column should raise an error if the column is absent.
     """
     with pytest.raises(ValueError):
         check_column(dataframe, ["b"])
+
+
+@pytest.mark.utils
+def test_check_column_excludes(dataframe):
+    """
+    check_column should return if column is absent and present is False
+    """
+    assert check_column(dataframe, ["b"], present=False) is None
+
+
+@pytest.mark.utils
+def test_check_column_absent_column_excludes(dataframe):
+    """
+    check_column should raise an error if the column is absent and present is
+    False
+    """
+    with pytest.raises(ValueError):
+        check_column(dataframe, ["a"], present=False)